Move formatDate method to date-util
