+++ src/ggrc-client/js/components/last-comment/last-comment_spec.js
@@ -8,7 +8,7 @@ import {getComponentVM} from '../../../js_specs/spec_helpers';
 import * as Utils from '../../plugins/utils/acl-utils.js';
 import RefreshQueue from '../../models/refresh_queue';
 import {COMMENT_CREATED} from '../../events/eventTypes';
-import {formatDate} from '../../plugins/ggrc_utils';
+import {formatDate} from '../../plugins/utils/date-util';
 
 describe('last-comment component', () => {
   let vm;
